requirements.txt
cloudpickle==0.5.3
cycler==0.10.0
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Table==0.5.0
html5lib==0.9999999
mysql-connector==2.1.6
nbconvert==5.3.1
nbformat==4.4.0
numpy==1.14.3
numpydoc==0.8.0
pandas==0.23.0
pickleshare==0.7.4
python-dateutil==2.7.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.1.0
sklearn==0.0
urllib3==1.22